+++ examples/cart-server-session/src/utils/cart.js
@@ -1,5 +1,6 @@
-import {gql} from '@shopify/hydrogen';
 import {flattenConnection} from '@shopify/hydrogen';
+import {graphqlClient} from '~/utils/graphqlClient';
+
 import {
   CART_GET_QUERY,
   CART_CREATE_MUTATION,
@@ -8,7 +9,6 @@ import {
 } from '~/graphql';
 
 export async function getCart({id}) {
-  console.log('getCart:getting cart with id', id);
   try {
     const result = await graphqlClient({
       query: CART_GET_QUERY,
@@ -33,7 +33,6 @@ export async function getCart({id}) {
 }
 
 export async function createCart(input) {
-  console.log('createCart:creating a cart with', input);
   const {data, errors} = await graphqlClient({
     query: CART_CREATE_MUTATION,
     variables: {input},
@@ -51,7 +50,6 @@ export async function createCart(input) {
 }
 
 export async function cartLinesAdd(input) {
-  console.log('cartLinesAdd:Adding lines to the cart', input);
   const {data, errors} = await graphqlClient({
     query: CART_LINES_ADD_MUTATION,
     variables: input,
@@ -69,7 +67,6 @@ export async function cartLinesAdd(input) {
 }
 
 export async function cartLinesRemove(input) {
-  console.log('cartLinesRemove:Removing lines from the cart', input);
   const {data, errors} = await graphqlClient({
     query: CART_LINES_REMOVE_MUTATION,
     variables: input,
@@ -84,58 +81,4 @@ export async function cartLinesRemove(input) {
     ...data.query.cart,
     lines: flattenConnection(data.query.cart.lines),
   };
-}
-
-/**
- * A basic Admin API fetch-based client.
- * @param {gql} query - GraphQL query
- * @param {object} variables - GraphQL variables
- * @returns {object} - {error: [], data: object}
- */
-async function graphqlClient(
-  {query, variables} = {id: '', query: null, variables: {}}
-) {
-  if (!query) {
-    throw new Error('Must provide a `query` to the admin client');
-  }
-
-  const endpoint = `https://hydrogen-preview.myshopify.com/api/2022-07/graphql.json`;
-  const options = {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'X-Shopify-Storefront-Access-Token': '3b580e70970c4528da70c98e097c2fa0',
-    },
-    body: JSON.stringify({
-      query,
-      variables,
-    }),
-  };
-
-  const request = await fetch(endpoint, options);
-
-  if (!request.ok) {
-    throw new Error(
-      `graphql api request not ok ${request.status} ${request.statusText}`
-    );
-  }
-
-  try {
-    const response = await request.json();
-
-    if (response?.errors?.length) {
-      throw new Error(response.errors[0].message);
-    }
-
-    return {
-      error: null,
-      data: response.data,
-    };
-  } catch (error) {
-    console.log('graphqlClient error', error);
-    return {
-      error: error.message,
-      data: null,
-    };
-  }
 }
